# apps/api/openapi/spec.yaml
openapi: 3.0.3
info:
  title: lomi. API
  version: 1.1.0
  description: Payment processing API for francophone West African businesses.
  contact:
    name: lomi. Support
    email: hello@lomi.africa
    url: https://lomi.africa
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.lomi.africa/v1
    description: Production server
  - url: https://sandbox.api.lomi.africa/v1
    description: Sandbox server
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: API key for authentication. Get yours from the lomi. dashboard.
  schemas:
...
    customers:
      type: object
      description: customers resource object
      properties:
        address:
          type: string
        city:
          type: string
        country:
          type: string
        created_at:
          type: string
          readOnly: true
          format: date-time
          description: ISO 8601 datetime
        created_by:
          type: string
          readOnly: true
        customer_id:
          type: string
          readOnly: true
          format: uuid
          description: Unique identifier (UUID format)
        deleted_at:
          type: string
          readOnly: true
          format: date-time
          description: ISO 8601 datetime
        email:
          type: string
          format: email
          description: Email address
        environment:
          type: string
        is_business:
          type: boolean
        is_deleted:
          type: boolean
          description: Soft deletion flag
        metadata:
          type: object
          description: Set of key-value pairs for storing additional information
        name:
          type: string
        organization_id:
          type: string
          format: uuid
          description: Unique identifier (UUID format)
        phone_number:
          type: string
        postal_code:
          type: string
        spi_alias_mbno:
          type: string
        spi_alias_shid:
          type: string
        spi_primary_alias:
          type: string
        updated_at:
          type: string
          readOnly: true
          format: date-time
          description: ISO 8601 datetime
        whatsapp_number:
          type: string
    customers_create:
      type: object
      description: Request body for creating a customers object. System-managed fields like `created_at`, `organization_id`, and IDs are automatically set.
      required: []
      properties:
        address:
          type: string
        city:
          type: string
        country:
          type: string
        email:
          type: string
          format: email
          description: Email address
        is_business:
          type: boolean
        is_deleted:
          type: boolean
          description: Soft deletion flag
        metadata:
          type: object
          description: Set of key-value pairs for storing additional information
        name:
          type: string
        phone_number:
          type: string
        postal_code:
          type: string
        spi_alias_mbno:
          type: string
        spi_alias_shid:
          type: string
        spi_primary_alias:
          type: string
        whatsapp_number:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
  /customers:
    get:
      operationId: listCustomers
      summary: List customers
      description: Customer management - create and manage customer profiles
      tags:
        - Customers
      security:
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of items to return (1-100)
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of items to skip for pagination
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: sort
          in: query
          description: 'Sort order. Format: `field:direction` (e.g., `created_at:desc`)'
          schema:
            type: string
            example: created_at:desc
      responses:
        '200':
          description: Successful response with paginated data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/customers'
                  pagination:
                    type: object
                    properties:
                      limit:
                        type: integer
                        description: Number of items per page
                      offset:
                        type: integer
                        description: Number of items skipped
                      total:
                        type: integer
                        description: Total number of items available
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createCustomer
      summary: Create customer
      description: Customer management - create and manage customer profiles
      tags:
        - Customers
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/customers_create'
            examples:
              default:
                summary: Sample customer creation
                value:
                  name: Aminata Diallo
                  email: aminata.diallo@example.com
                  phone_number: '+225021234567'
                  metadata:
                    customer_segment: premium
                    referral_source: instagram
      responses:
        '201':
          description: Customer successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/customers'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /customers/{customer_id}:
    get:
      operationId: retrieveCustomer
      summary: Retrieve customer
      description: Retrieve a specific customer by its unique identifier.
      tags:
        - Customers
      security:
        - ApiKeyAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          description: Unique identifier for the customer
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customer retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/customers'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateCustomer
      summary: Update customer
      description: Update a specific customer. Only provided fields will be updated.
      tags:
        - Customers
      security:
        - ApiKeyAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          description: Unique identifier for the customer
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/customers_update'
            examples:
              default:
                summary: Sample customer update
                value:
                  name: Aminata Diallo-Kane
                  phone_number: '+225779876543'
                  metadata:
                    customer_segment: vip
                    notes: Upgraded to VIP tier
      responses:
        '200':
          description: Customer successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/customers'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteCustomer
      summary: Delete customer
      description: Delete a specific customer. This action cannot be undone.
      tags:
        - Customers
      security:
        - ApiKeyAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          description: Unique identifier for the customer
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Customer successfully deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /transactions:
    get:
      operationId: listTransactions
      summary: List transactions
      description: Transaction history - view completed and pending transactions
      tags:
        - Transactions
      security:
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of items to return (1-100)
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of items to skip for pagination
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: sort
          in: query
          description: 'Sort order. Format: `field:direction` (e.g., `created_at:desc`)'
          schema:
            type: string
            example: created_at:desc
      responses:
        '200':
          description: Successful response with paginated data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/transactions'
                  pagination:
                    type: object
                    properties:
                      limit:
                        type: integer
                        description: Number of items per page
                      offset:
                        type: integer
                        description: Number of items skipped
                      total:
                        type: integer
                        description: Total number of items available
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /transactions/{transaction_id}:
    get:
      operationId: retrieveTransaction
      summary: Retrieve transaction
      description: Retrieve a specific transaction by its unique identifier.
      tags:
        - Transactions
      security:
        - ApiKeyAuth: []
      parameters:
        - name: transaction_id
          in: path
          required: true
          description: Unique identifier for the transaction
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/transactions'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /refunds:
    get:
      operationId: listRefunds
      summary: List refunds
      description: Refund management - process and track refunds
      tags:
        - Refunds
      security:
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of items to return (1-100)
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of items to skip for pagination
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: sort
          in: query
          description: 'Sort order. Format: `field:direction` (e.g., `created_at:desc`)'
          schema:
            type: string
            example: created_at:desc
      responses:
        '200':
          description: Successful response with paginated data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/refunds'
                  pagination:
                    type: object
                    properties:
                      limit:
                        type: integer
                        description: Number of items per page
                      offset:
                        type: integer
                        description: Number of items skipped
                      total:
                        type: integer
                        description: Total number of items available
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createRefund
      summary: Create refund
      description: Refund management - process and track refunds
      tags:
        - Refunds
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/refunds_create'
            examples:
              default:
                summary: Sample refund creation
                value:
                  name: Sample refund
                  description: Example refund object
      responses:
        '201':
          description: Refund successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/refunds'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /refunds/{refund_id}:
    get:
      operationId: retrieveRefund
      summary: Retrieve refund
      description: Retrieve a specific refund by its unique identifier.
      tags:
        - Refunds
      security:
        - ApiKeyAuth: []
      parameters:
        - name: refund_id
          in: path
          required: true
          description: Unique identifier for the refund
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Refund retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/refunds'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateRefund
      summary: Update refund
      description: Update a specific refund. Only provided fields will be updated.
      tags:
        - Refunds
      security:
        - ApiKeyAuth: []
      parameters:
        - name: refund_id
          in: path
          required: true
          description: Unique identifier for the refund
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/refunds_update'
            examples:
              default:
                summary: Sample refund update
                value:
                  metadata:
                    updated_at: '2025-11-21T13:32:16.239Z'
                    updated_reason: Administrative update
      responses:
        '200':
          description: Refund successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/refunds'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteRefund
      summary: Delete refund
      description: Delete a specific refund. This action cannot be undone.
      tags:
        - Refunds
      security:
        - ApiKeyAuth: []
      parameters:
        - name: refund_id
          in: path
          required: true
          description: Unique identifier for the refund
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Refund successfully deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /products:
    get:
      operationId: listProducts
      summary: List products
      description: Product catalog - manage one-time and recurring products
      tags:
        - Products
      security:
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of items to return (1-100)
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of items to skip for pagination
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: sort
          in: query
          description: 'Sort order. Format: `field:direction` (e.g., `created_at:desc`)'
          schema:
            type: string
            example: created_at:desc
      responses:
        '200':
          description: Successful response with paginated data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/products'
                  pagination:
                    type: object
                    properties:
                      limit:
                        type: integer
                        description: Number of items per page
                      offset:
                        type: integer
                        description: Number of items skipped
                      total:
                        type: integer
                        description: Total number of items available
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createProduct
      summary: Create product
      description: Product catalog - manage one-time and recurring products
      tags:
        - Products
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/products_create'
            examples:
              default:
                summary: Sample product creation
                value:
                  name: Premium Subscription
                  description: Monthly premium access with all features
                  product_type: recurring
                  is_active: true
                  metadata:
                    features:
                      - analytics
                      - api_access
                      - priority_support
      responses:
        '201':
          description: Product successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/products'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...